move methods to DType
