+++ vortex-dtype/src/field.rs
@@ -9,9 +9,6 @@ use std::fmt::{Display, Formatter};
 use std::sync::Arc;
 
 use itertools::Itertools;
-use vortex_error::{VortexResult, vortex_bail};
-
-use crate::DType;
 
 /// Selects a nested type within either a struct or a list.
 #[derive(Clone, Debug, PartialEq, Eq, Hash)]
@@ -73,7 +70,7 @@ impl Field {
 /// use vortex_dtype::*;
 ///
 /// let dtype_i32 = DType::Primitive(PType::I32, Nullability::NonNullable);
-/// assert_eq!(dtype_i32, FieldPath::root().resolve(dtype_i32.clone()).unwrap());
+/// assert_eq!(dtype_i32, dtype_i32.clone().resolve(&FieldPath::root()).unwrap());
 /// ```
 ///
 // TODO(ngates): we should probably reverse the path. Or better yet, store a Arc<[Field]> along
@@ -126,62 +123,6 @@ impl FieldPath {
         self.0 = self.0.iter().skip(1).cloned().collect();
         Some(self)
     }
-
-    /// The dtype, within the given type, to which this field path refers.
-    ///
-    /// Note that a nullable DType may contain a non-nullable DType. This function returns the
-    /// literal nullability of the child.
-    ///
-    /// # Examples
-    ///
-    /// Extract the type of the "b" field from `struct{a: list(struct{b: u32})?}`:
-    ///
-    /// ```
-    /// use std::sync::Arc;
-    ///
-    /// use vortex_dtype::*;
-    /// use vortex_dtype::Nullability::*;
-    ///
-    /// let dtype = DType::Struct(
-    ///     Arc::new(StructFields::from_iter([(
-    ///         "a",
-    ///         DType::List(
-    ///             Arc::new(DType::Struct(
-    ///                 Arc::new(StructFields::from_iter([(
-    ///                     "b",
-    ///                     DType::Primitive(PType::U32, NonNullable),
-    ///                 )])),
-    ///                 NonNullable,
-    ///             )),
-    ///             Nullable,
-    ///         ),
-    ///     )])),
-    ///     NonNullable,
-    /// );
-    ///
-    /// let path = FieldPath::from(vec![Field::from("a"), Field::ElementType, Field::from("b")]);
-    /// let resolved = path.resolve(dtype).unwrap();
-    /// assert_eq!(resolved, DType::Primitive(PType::U32, NonNullable));
-    /// ```
-    pub fn resolve(&self, mut dtype: DType) -> VortexResult<DType> {
-        for field in &self.0 {
-            dtype = match (dtype, field) {
-                (DType::Struct(fields, _), Field::Name(name)) => fields.field(name)?,
-                (DType::List(element_dtype, _), Field::ElementType) => DType::clone(&element_dtype),
-                (other, f) => {
-                    vortex_bail!("FieldPath: invalid index {:?} for DType {:?}", f, other)
-                }
-            }
-        }
-
-        Ok(dtype)
-    }
-
-    /// Does the field referenced by the field path exist in the given dtype?
-    pub fn exists(&self, dtype: DType) -> bool {
-        // Indexing a struct type always allocates anyway.
-        self.resolve(dtype).is_ok()
-    }
 }
 
 impl FromIterator<Field> for FieldPath {
@@ -211,8 +152,6 @@ impl Display for FieldPath {
 #[cfg(test)]
 mod tests {
     use super::*;
-    use crate::Nullability::*;
-    use crate::{DType, PType, StructFields};
 
     #[test]
     fn test_field_path() {
@@ -228,131 +167,5 @@ mod tests {
         let vec_path = FieldPath::from(fields);
         assert_eq!(vec_path.to_string(), "$A.$B.$C");
         assert_eq!(path, vec_path);
-    }
-
-    #[test]
-    fn nested_field_single_level() {
-        let a_type = DType::Primitive(PType::I32, NonNullable);
-        let dtype = DType::Struct(
-            Arc::from(StructFields::from_iter([
-                ("a", a_type.clone()),
-                ("b", DType::Bool(Nullable)),
-            ])),
-            NonNullable,
-        );
-        let path = FieldPath::from_name("a");
-        assert_eq!(a_type, path.resolve(dtype.clone()).unwrap());
-        assert!(path.exists(dtype));
-    }
-
-    #[test]
-    fn nested_field_two_level() {
-        let inner = DType::Struct(
-            Arc::new(StructFields::from_iter([
-                ("inner_a", DType::Primitive(PType::U8, NonNullable)),
-                ("inner_b", DType::Bool(Nullable)),
-            ])),
-            NonNullable,
-        );
-
-        let outer = DType::Struct(
-            Arc::from(StructFields::from_iter([
-                ("outer_a", DType::Bool(NonNullable)),
-                ("outer_b", inner),
-            ])),
-            NonNullable,
-        );
-
-        let path = FieldPath::from_name("outer_b").push("inner_a");
-        let dtype = path.resolve(outer.clone()).unwrap();
-
-        assert_eq!(dtype, DType::Primitive(PType::U8, NonNullable));
-        assert!(path.exists(outer));
-    }
-
-    #[test]
-    fn nested_field_deep_nested() {
-        let level4 = DType::Struct(
-            Arc::new(StructFields::from_iter([(
-                "c",
-                DType::Primitive(PType::F64, Nullable),
-            )])),
-            NonNullable,
-        );
-
-        let level3 = DType::List(Arc::from(level4), Nullable);
-
-        let level2 = DType::Struct(
-            Arc::new(StructFields::from_iter([("b", level3)])),
-            NonNullable,
-        );
-
-        let level1 = DType::Struct(
-            Arc::from(StructFields::from_iter([("a", level2)])),
-            NonNullable,
-        );
-
-        let path = FieldPath::from_name("a")
-            .push("b")
-            .push(Field::ElementType)
-            .push("c");
-        let dtype = path.resolve(level1.clone()).unwrap();
-
-        assert_eq!(dtype, DType::Primitive(PType::F64, Nullable));
-        assert!(path.exists(level1.clone()));
-
-        let path = FieldPath::from_name("a")
-            .push("b")
-            .push("c")
-            .push(Field::ElementType);
-        assert!(path.resolve(level1.clone()).is_err());
-        assert!(!path.exists(level1.clone()));
-
-        let path = FieldPath::from_name("a")
-            .push(Field::ElementType)
-            .push("b")
-            .push("c");
-        assert!(path.resolve(level1.clone()).is_err());
-        assert!(!path.exists(level1.clone()));
-
-        let path = FieldPath::from_name(Field::ElementType)
-            .push("a")
-            .push("b")
-            .push("c");
-        assert!(path.resolve(level1.clone()).is_err());
-        assert!(!path.exists(level1.clone()));
-    }
-
-    #[test]
-    fn nested_field_not_found() {
-        let dtype = DType::Struct(
-            Arc::from(StructFields::from_iter([("a", DType::Bool(NonNullable))])),
-            NonNullable,
-        );
-        let path = FieldPath::from_name("b");
-        assert!(path.resolve(dtype.clone()).is_err());
-        assert!(!path.exists(dtype.clone()));
-
-        let path = FieldPath::from(Field::ElementType);
-        assert!(path.resolve(dtype.clone()).is_err());
-        assert!(!path.exists(dtype.clone()));
-    }
-
-    #[test]
-    fn nested_field_non_struct_intermediate() {
-        let dtype = DType::Struct(
-            Arc::from(StructFields::from_iter([(
-                "a",
-                DType::Primitive(PType::I32, NonNullable),
-            )])),
-            NonNullable,
-        );
-        let path = FieldPath::from_name("a").push("b");
-        assert!(path.resolve(dtype.clone()).is_err());
-        assert!(!path.exists(dtype.clone()));
-
-        let path = FieldPath::from_name("a").push(Field::ElementType);
-        assert!(path.resolve(dtype.clone()).is_err());
-        assert!(!path.exists(dtype.clone()));
     }
 }
